\name{bc_classes}
\alias{bc_classes}
\docType{data}
\title{Breast cancer microarray example dataset - class labels}
\description{
The class labels for the breast cancer microarray example dataset
}
\usage{data(bc_classes)}
\details{
  This vector contains the class labels for the breast cancer microarray
  example dataset (see bc\_dat help page)
}
\source{
  The original data set is publicly available at
  \url{http://www.ebi.ac.uk/microarray-as/ae}
  under the accession number: E-TABM-576
}
\keyword{datasets}
