\name{bc_dat}
\alias{bc_dat}
\docType{data}
\title{ Breast cancer microarray example dataset}
\description{
   Example of a 3D input structure to illustrate VRMLGen's visualization methods
}
\usage{data(bc_dat)}
%  The format is:
% num [1:76, 1:3] -0.3723 -0.3263 0.0693 0.1001 -0.9049 ...
% - attr(*, "dimnames")=List of 2
%  ..$ : chr [1:76] "V2" "V3" "V4" "V5" ...
%  ..$ : NULL
\format{
}
\details{
  This dataset was obtained by applying an Independent Component Analysis
  to a public breast cancer microarray data set. The resulting three
  independent component vectors can be visualized in 3D and are
  used to illustrate some of VRMLGen's features.
}
\source{
  The original data set is publicly available at
  \url{http://www.ebi.ac.uk/microarray-as/ae}
  under the accession number: E-TABM-576
}
\references{  
}
\examples{
}
\keyword{datasets}
