\name{c60coords}
\alias{c60coords}
\docType{data}
\title{ Atom coordinates of a C60 molecule - example dataset}
\description{
  Example of a 3D input structure to illustrate VRMLGen's visualization methods
}
\usage{data(c60coords)}
%The format is:
% num [1:60, 1:3] -1 -1 1 1 0 ...
\details{
  This dataset contains the 3D-coordinates of a C60 molecule (known as fullerene)
  used as example input to demonstrate VRMLGen's visualization features.
}
\references{
  The molecular structure was created with the open-source molecular modeling 
  sofware BALL, see www.ballview.org.
}
\keyword{datasets}
