\name{theo_edges}
\alias{theo_edges}
\docType{data}
\title{ Theobromine molecular structure example dataset}
\description{
  Example of a 3D input structure for the mesh3d-function (edges given as 3D-vertex indices).
}
\usage{data(theo_edges)}
%The format is:
% num [1:8662, 1:4] 1 2 3 4 5 6 7 8 9 10 ...
\format{ 
}
\details{
  This dataset is an example input for the mesh3d-function and contains information
  on the polygonal faces of a molecular structure (theobromine, a molecule similar to caffeine).
}
\source{
  The molecular structure was created with the open-source molecular modeling 
  sofware BALL, see www.ballview.org.  
}
\references{  
}
\examples{
}
\keyword{datasets}
