\name{theo_vertices}
\alias{theo_vertices}
\docType{data}
\title{ Theobromine molecular structure example dataset }
\description{
  Example of a 3D input structure for the mesh3d-function (3D-vertex coordinates).
}
% The format is:
% num [1:8966, 1:3] 0.904 0.895 0.882 0.865 0.847 ...
\usage{data(theo_vertices)}
\details{
  This dataset contains the 3D-vertices of a molecular structure used 
  as example input for the mesh3d-function. The molecular structure
  represents theobromine, a molecule similar to caffeine.
}
\source{  
  The molecular structure was created with the open-source molecular modeling 
  sofware BALL, see www.ballview.org.  
}
\keyword{datasets}
