\name{vrmlgen-internal}
\title{Internal vrmlgen objects}
\alias{.hnorm}
\alias{.vdense}
\alias{.vdensecalc}
\alias{.colorselection}
\alias{.lbar}
\alias{.lcloud}
\alias{.lmesh}
\alias{.vbar}
\alias{.vcloud}
\alias{.vmesh}
\alias{.lg3dpoints}
\alias{.lg3dtext}
\alias{.vrmlpoints}
\alias{.vrmltext}
\alias{.multquaternion}
\alias{.quaternion2rot}
\alias{.rot2quaternion}
\description{Internal vrmlgen objects.}
\details{These are not to be called by the user.}
\keyword{internal}